final gemini chatbot/requirements.txt
openai
streamlit
python-dotenv
google-generativeai